Move magic link api tests to their own files
